# ptak-expo-backend/swagger.yaml
openapi: 3.0.3
info:
  title: PTAK EXPO Backend API
  version: 1.0.0
  description: REST API for exhibitor and exhibition management
servers:
  - url: http://localhost:3001
paths:
...
  /api/v1/exhibitors:
    get:
      summary: List exhibitors (admin)
      tags: [Exhibitors]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of exhibitors
    post:
      summary: Create exhibitor (admin)
      tags: [Exhibitors]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nip: { type: string }
                companyName: { type: string }
                address: { type: string }
                postalCode: { type: string }
                city: { type: string }
                contactPerson: { type: string }
                contactRole: { type: string }
                phone: { type: string }
                email: { type: string }
                password: { type: string }
                exhibitionId: { type: integer, nullable: true }
      responses:
        '201': { description: Created }
  /api/v1/exhibitors/{id}:
    get:
      summary: Get exhibitor detail (admin)
      tags: [Exhibitors]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Exhibitor }
    delete:
      summary: Delete exhibitor (admin)
      tags: [Exhibitors]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Deleted }
  /api/v1/exhibitors/{id}/assign-event:
    post:
      summary: Assign exhibitor to exhibition (admin)
      tags: [Exhibitors]
      security: [ { bearerAuth: [] } ]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                exhibitionId: { type: integer }
      responses:
        '200': { description: Assigned }
  /api/v1/exhibitors/{id}/assign-event/{exhibitionId}:
    delete:
      summary: Unassign exhibitor from exhibition (admin)
      tags: [Exhibitors]
      security: [ { bearerAuth: [] } ]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
        - in: path
          name: exhibitionId
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Unassigned }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT